move prepare_cpp_array to buffer handling
